+++ src/config.c
@@ -132,6 +132,8 @@ const config default_cfg = {
   .hide_mouse = true,
   .elastic_mouse = false,
   .lines_per_notch = 0,
+  .mouse_pointer = W("ibeam"),
+  .appmouse_pointer = W("arrow"),
   // Selection
   .input_clears_selection = true,
   .copy_on_select = true,
@@ -409,6 +411,8 @@ options[] = {
   {"HideMouse", OPT_BOOL, offcfg(hide_mouse)},
   {"ElasticMouse", OPT_BOOL, offcfg(elastic_mouse)},
   {"LinesPerMouseWheelNotch", OPT_INT, offcfg(lines_per_notch)},
+  {"MousePointer", OPT_WSTRING, offcfg(mouse_pointer)},
+  {"AppMousePointer", OPT_WSTRING, offcfg(appmouse_pointer)},
 
   // Selection
   {"ClearSelectionOnInput", OPT_BOOL, offcfg(input_clears_selection)},