+++ src/termout.c
@@ -3982,6 +3982,11 @@ do_cmd(void)
           }
         }
       }
+    when 22: {  // set mouse pointer style
+      wchar * ps = cs__mbstowcs(s);
+      set_cursor_style(term.mouse_mode || term.locator_1_enabled, ps);
+      free(ps);
+    }
     when 7750:
       set_arg_option("Emojis", strdup(s));
       clear_emoji_data();